make sessions optional if params used are overriden
