+++ packages/core/src/__types__/error.ts
@@ -1,0 +1,26 @@
+/**@internal */
+export class TurnkeyRequestError extends Error {
+  details: any[] | null;
+  code: number;
+
+  constructor(input: GrpcStatus) {
+    let turnkeyErrorMessage = `Turnkey error ${input.code}: ${input.message}`;
+
+    if (input.details != null) {
+      turnkeyErrorMessage += ` (Details: ${JSON.stringify(input.details)})`;
+    }
+
+    super(turnkeyErrorMessage);
+
+    this.name = "TurnkeyRequestError";
+    this.details = input.details ?? null;
+    this.code = input.code;
+  }
+}
+
+/**@internal */
+export type GrpcStatus = {
+  message: string;
+  code: number;
+  details: unknown[] | null;
+};
