+++ packages/core/src/__types__/export.ts
@@ -1,0 +1,10 @@
+/** @internal */
+export interface IframeClientParams {
+  iframeContainer: HTMLElement | null | undefined;
+  iframeUrl: string;
+  iframeElementId?: string;
+  dangerouslyOverrideIframeKeyTtl?: number;
+}
+
+/** @internal */
+export type ExportBundle = string;
