+++ packages/core/src/index.ts
@@ -17,7 +17,7 @@ export { TurnkeyClient, type TurnkeyClientMethods } from "./__clients__/core";
 export { type TurnkeySDKClientBase } from "./__generated__/sdk-client-base";
 
 // Export all types and values from __types__/base
-export * from "./__types__/base";
+export * from "./__types__/index";
 
 /**@internal */
 export {
