+++ packages/react-wallet-kit/src/index.ts
@@ -2,5 +2,5 @@
 import "./index.css";
 export * from "./providers";
 /** @internal */
-export * from "@turnkey/core/dist/__types__/base";
+export * from "@turnkey/core/src/__types__";
 export * from "./types/base";
